[update] added documentation for routes in iam, engines, experiments and models

+++ back-end/src/routers/iam.py
@@ -92,9 +92,7 @@ async def delete_user(
     try:
         async with await mongo_client.start_session() as session:
             async with session.start_transaction():
-                await db["users"].delete_many(
-                    {"userId": {"$in": userid.users}}
-                )
+                await db["users"].delete_many({"userId": {"$in": userid.users}})
     except Exception as err:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND, detail="User not found"
@@ -110,6 +108,17 @@ async def update_user(
     user: UserInsert,
     db: Tuple[AsyncIOMotorDatabase, AsyncIOMotorClient] = Depends(get_db),
 ):
+    """Edit a single user from the database
+
+    Args:
+        user (UserInsert): User to edit
+        db (Tuple[AsyncIOMotorDatabase, AsyncIOMotorClient], optional): MongoDB connection.
+            Defaults to Depends(get_db).
+
+    Raises:
+        HTTPException: 404 if user not found
+        HTTPException: 422 if value error occurs
+    """
     db, mongo_client = db
     try:
         user.password = get_password_hash(user.password.get_secret_value())
@@ -147,6 +156,18 @@ async def update_many_user(
     user: UsersEdit,
     db: Tuple[AsyncIOMotorDatabase, AsyncIOMotorClient] = Depends(get_db),
 ):
+    """Bulk edit many users' privileges in the database
+
+    Args:
+        user (UsersEdit): Users to edit
+        db (Tuple[AsyncIOMotorDatabase, AsyncIOMotorClient], optional): MongoDB connection.
+            Defaults to Depends(get_db).
+
+    Raises:
+        HTTPException: 400 if privilege to change to was not defined
+        HTTPException: 404 if users not found
+        HTTPException: 422 if value error occurs
+    """
     db, mongo_client = db
     try:
         if user.priv is None:
